requirements.txt
matplotlib
cupy-cuda12x
openpyxl
tqdm
pillow
opencv-python
XlsxWriter
cellpose[gui]
